Delete filteredOffers, sortedOffers, refactor
