+++ project/src/pages/private-route/private-route.tsx
@@ -13,7 +13,7 @@ function PrivateRoute(props: PrivateRouteProps): JSX.Element {
   const authorizationStatus = useAppSelector((state) => state.authorizationStatus);
 
   return (
-    authorizationStatus === AuthorizationStatus.Auth
+    authorizationStatus !== AuthorizationStatus.NoAuth
       ? children
       : <Navigate to={AppRoute.Login} />
   );
