+++ project/src/store/reducer.ts
@@ -1,8 +1,6 @@
 import {createReducer} from '@reduxjs/toolkit';
 import {changeFilterType,
-  filterCity,
   changeSortType,
-  sortOffers,
   selectOffer,
   loadOffers,
   loadUserData,
@@ -14,7 +12,6 @@ import {changeFilterType,
   requireAuthorization,
   setDownloadError,
   loadFavoriteOffers,
-  changeFavoriteStatusOffer,
 } from './action';
 import {Offer} from '../types/offer';
 import {Review} from '../types/review';
@@ -22,22 +19,18 @@ import {City} from '../types/city';
 import {SortType} from '../types/sort-type';
 import {MAP_CITIES} from '../mocks/map-cities';
 import {DEFAULT_FILTER_TYPE, DEFAULT_SORT_TYPE} from '../const';
-import {getFilterOffers,
+import {
   getFilterCity,
   getHoverOffer,
-  getSortOffers,
   getSortReviews,
-  changeOffers,
-} from './utils';
+} from '../utils';
 import {AuthorizationStatus} from '../const';
 import {UserData} from '../types/user-data';
 
 type InitialState = {
   offers: Offer[],
   filterType: string,
-  filteredOffers: Offer[],
   sortType: SortType,
-  sortedOffers: Offer[],
   selectedOffer?: Offer,
   propertyOffer?: Offer,
   nearbyOffers?: Offer[],
@@ -53,9 +46,7 @@ type InitialState = {
 const initialState: InitialState = {
   offers: [],
   filterType: DEFAULT_FILTER_TYPE,
-  filteredOffers: [],
   sortType: DEFAULT_SORT_TYPE,
-  sortedOffers: [],
   favoriteOffers: [],
   mapCity: getFilterCity(MAP_CITIES, DEFAULT_FILTER_TYPE),
   authorizationStatus : AuthorizationStatus.Unknown,
@@ -70,24 +61,13 @@ const reducer = createReducer(initialState, (builder) => {
 
       state.filterType = action.payload.city;
     })
-    .addCase(filterCity, (state) => {
-
-      state.filteredOffers = getFilterOffers(state.offers, state.filterType);
-      state.mapCity = getFilterCity(MAP_CITIES, state.filterType);
-      state.sortedOffers = state.filteredOffers.slice();
-      state.sortType = DEFAULT_SORT_TYPE;
-    })
     .addCase(changeSortType, (state, action) => {
 
       state.sortType = action.payload.sortType;
     })
-    .addCase(sortOffers, (state) => {
-
-      state.sortedOffers = getSortOffers(state.filteredOffers, state.sortType.type);
-    })
     .addCase(selectOffer, (state, action) => {
 
-      state.selectedOffer = getHoverOffer(state.filteredOffers, action.payload.selectedOfferId);
+      state.selectedOffer = getHoverOffer(state.offers, action.payload.selectedOfferId);
     })
     .addCase(loadOffers, (state, action) => {
 
@@ -112,12 +92,6 @@ const reducer = createReducer(initialState, (builder) => {
     .addCase(loadFavoriteOffers, (state, action) => {
 
       state.favoriteOffers = action.payload;
-    })
-    .addCase(changeFavoriteStatusOffer, (state, action) => {
-
-      state.offers = changeOffers(state.offers, action.payload);
-      state.filteredOffers = changeOffers(state.filteredOffers, action.payload);
-      state.sortedOffers = changeOffers(state.sortedOffers, action.payload);
     })
     .addCase(setDataLoadedStatus, (state, action) => {
